Move /market page to homepage
